% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NBBDesigns.R
\name{anlys}
\alias{anlys}
\title{Analysis of  data}
\usage{
anlys(data, yield, block, treatment, left_neighbour, right_neighbour)
}
\arguments{
\item{data}{The data file should be in csv format.}

\item{yield}{Provide yield column name (eg. "Yield").}

\item{block}{Provide the block column name (eg. "block").}

\item{treatment}{Provide the treatment column name (eg. "Trt").}

\item{left_neighbour}{Provide the left_neighbour column name (eg. "Left Neighbour").}

\item{right_neighbour}{Provide the right_neighbour column name (eg. "Right Neighbour").}
}
\value{
It provides the ANOVA table.
}
\description{
This function provides the Analysis of Variance (Type III) of the data generated from experiments conducted using a neighbour balanced/partially neighbour balanced block design.
}
\examples{
library(NBBDesigns)
sample_data

anlys(
  data = sample_data,
  yield = "yield",
  block = "block",
  treatment = "treatment",
  left_neighbour = "left_neighbour",
  right_neighbour = "right_neighbour"
)
}
