% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NBBDesigns.R
\name{nbbd2}
\alias{nbbd2}
\title{NBB Design  of Type 2 (NBBD 2)}
\usage{
nbbd2(v)
}
\arguments{
\item{v}{Number of treatments (v is odd number but not a power of a prime number)}
}
\value{
It generates neighbour balanced block designs for a given number of treatments (v>3),  v is odd number but not a power of a prime number.
}
\description{
When v is odd number but not a power of a prime number (v<= 100) , the function generates a class of neighbour balanced block designs for a given number of treatments. This design listed as type 1 design (Azais, 1993) which is constructed using the method of Rees (1967) and Azais (1987). It also gives the parameters of the design, information matrix for estimating the contrast pertaining to direct and neighbour effects (both left and right) of the treatments.
}
\examples{
library(NBBDesigns)
nbbd2(7)
}
\references{
Azais, J.M., Bailey, R.A. and Monod, H. (1993)<DOI: 10.2307/2532269>."A catalogue of efficient neighbour designs with border plots".
}
