% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NBBDesigns.R
\name{anlys}
\alias{anlys}
\title{Analysis of  data}
\usage{
anlys(data)
}
\arguments{
\item{data}{The data file should be in csv format. The columns should be named as block,treatment, left_neighbour, right_neighbour and yield as given in the example data set.}
}
\value{
It provides the ANOVA table.
}
\description{
This function provides the Analysis of Variance (Type III) of the data generated from experiments conducted using a neighbour balanced/partially neighbour balanced block design.
}
\examples{
\dontrun{
library(NBBDesigns)
data<-file.choose()
data<-read.csv(data,header=TRUE,colClasses = c("factor","factor","factor","factor","numeric"))
fix(data)
anlys(data)
}
}
